# docs/swagger.yaml
basePath: /api
definitions:
  domain.Day:
    properties:
      fri:
        type: boolean
      id:
        type: integer
      mon:
        type: boolean
      sat:
        type: boolean
      sun:
        type: boolean
      thu:
        type: boolean
      time:
        type: string
      tue:
        type: boolean
      userID:
        type: integer
      wed:
        type: boolean
    type: object
  domain.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  domain.Favorites:
    properties:
      id:
        type: integer
      lessonID:
        type: integer
      userID:
        type: integer
    type: object
  domain.Lesson:
    properties:
      created_at:
        type: string
      description:
        type: string
      duration:
        type: number
      id:
        type: integer
      imageSrc:
        type: string
      name:
        type: string
      title:
        type: string
      updated_at:
        type: string
      videoSrc:
        type: string
    type: object
...
  domain.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /days:
    get:
      description: get day
      operationId: get-Day
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Day'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetDay
      tags:
      - Day
    post:
      consumes:
      - application/json
      description: create Day
      operationId: create-Day
      parameters:
      - description: Create Day
        in: body
        name: Day
        required: true
        schema:
          $ref: '#/definitions/domain.Day'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Day'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: CreateDay
      tags:
      - Day
  /days/update:
    post:
      consumes:
      - application/json
      description: update day
      operationId: update-day
      parameters:
      - description: update day
        in: body
        name: day
        required: true
        schema:
          $ref: '#/definitions/domain.Day'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Day'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: UpdateDay
      tags:
      - Day
  /favorites:
    get:
      description: get Favorites
      operationId: get-favorites
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Lesson'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetFavorites
      tags:
      - favorites
    post:
      consumes:
      - application/json
      description: create favorites
      operationId: create-favorites
      parameters:
      - description: Create favorites
        in: body
        name: favorites
        required: true
        schema:
          $ref: '#/definitions/domain.Favorites'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Favorites'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Createfavorites
      tags:
      - favorites
  /favorites/{lessonid}:
    delete:
      consumes:
      - application/json
      description: Delete favorites
      operationId: Delete-favorites
      parameters:
      - description: lesson ID
        in: path
        name: lessonid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete lesson from Favorites
      tags:
      - favorites
  /lessons:
    get:
      description: get all Lessons
      operationId: get-Lessons
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Lesson'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetLessons
      tags:
      - Lesson
    post:
      consumes:
      - application/json
      description: create Lesson
      operationId: create-Lesson
      parameters:
      - description: Create Lesson
        in: body
        name: Lesson
        required: true
        schema:
          $ref: '#/definitions/domain.Lesson'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Lesson'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: CreateLesson
      tags:
      - Lesson
  /lessons/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Lesson
      operationId: Delete-Lesson
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: DeleteLesson
      tags:
      - Lesson
    get:
      description: get Lesson by id
      operationId: get-Lesson
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Lesson'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetLesson
      tags:
      - Lesson
    post:
      consumes:
      - application/json
      description: update Lesson
      operationId: update-Lesson
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      - description: update Lesson
        in: body
        name: Lesson
        required: true
        schema:
          $ref: '#/definitions/domain.Lesson'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Lesson'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: UpdateLesson
      tags:
      - Lesson
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"